Improve exception handling

+++ peerprep-backend/src/main/java/com/peerprep/peerprepbackend/service/QuestionService.java
@@ -4,6 +4,7 @@ import com.peerprep.peerprepbackend.dto.request.CreateQuestionRequest;
 import com.peerprep.peerprepbackend.dto.response.QuestionOverview;
 import com.peerprep.peerprepbackend.dto.response.QuestionResponse;
 import com.peerprep.peerprepbackend.entity.Question;
+import com.peerprep.peerprepbackend.exception.QuestionNotFoundException;
 import com.peerprep.peerprepbackend.repository.QuestionRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -40,7 +41,7 @@ public class QuestionService {
 
     public QuestionResponse getQuestion(String id) {
         Question question = questionRepository.findById(id)
-                .orElseThrow(RuntimeException::new);
+                .orElseThrow(() -> new QuestionNotFoundException(id));
         return QuestionResponse.builder()
                 .id(question.getId())
                 .title(question.getTitle())
@@ -51,6 +52,9 @@ public class QuestionService {
     }
 
     public void deleteQuestion(String id) {
+        if (!questionRepository.existsById(id)) {
+            throw new QuestionNotFoundException(id);
+        }
         questionRepository.deleteById(id);
     }
 }
